+++ lib/src/theme/text_theme.freezed.dart
@@ -15,7 +15,7 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$WaveTextTheme {
 
- TextStyle get body;
+ TextStyle get h1; TextStyle get h2; TextStyle get h3; TextStyle get h4; TextStyle get large; TextStyle get body; TextStyle get small; TextStyle get button;
 /// Create a copy of WaveTextTheme
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -26,12 +26,12 @@ $WaveTextThemeCopyWith<WaveTextTheme> get copyWith => _$WaveTextThemeCopyWithImp
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is WaveTextTheme&&(identical(other.body, body) || other.body == body));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is WaveTextTheme&&(identical(other.h1, h1) || other.h1 == h1)&&(identical(other.h2, h2) || other.h2 == h2)&&(identical(other.h3, h3) || other.h3 == h3)&&(identical(other.h4, h4) || other.h4 == h4)&&(identical(other.large, large) || other.large == large)&&(identical(other.body, body) || other.body == body)&&(identical(other.small, small) || other.small == small)&&(identical(other.button, button) || other.button == button));
 }
 
 
 @override
-int get hashCode => Object.hash(runtimeType,body);
+int get hashCode => Object.hash(runtimeType,h1,h2,h3,h4,large,body,small,button);
 
 
 
@@ -42,7 +42,7 @@ abstract mixin class $WaveTextThemeCopyWith<$Res>  {
   factory $WaveTextThemeCopyWith(WaveTextTheme value, $Res Function(WaveTextTheme) _then) = _$WaveTextThemeCopyWithImpl;
 @useResult
 $Res call({
- TextStyle body
+ TextStyle h1, TextStyle h2, TextStyle h3, TextStyle h4, TextStyle large, TextStyle body, TextStyle small, TextStyle button
 });
 
 
@@ -59,9 +59,16 @@ class _$WaveTextThemeCopyWithImpl<$Res>
 
 /// Create a copy of WaveTextTheme
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? body = null,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? h1 = null,Object? h2 = null,Object? h3 = null,Object? h4 = null,Object? large = null,Object? body = null,Object? small = null,Object? button = null,}) {
   return _then(_self.copyWith(
-body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+h1: null == h1 ? _self.h1 : h1 // ignore: cast_nullable_to_non_nullable
+as TextStyle,h2: null == h2 ? _self.h2 : h2 // ignore: cast_nullable_to_non_nullable
+as TextStyle,h3: null == h3 ? _self.h3 : h3 // ignore: cast_nullable_to_non_nullable
+as TextStyle,h4: null == h4 ? _self.h4 : h4 // ignore: cast_nullable_to_non_nullable
+as TextStyle,large: null == large ? _self.large : large // ignore: cast_nullable_to_non_nullable
+as TextStyle,body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+as TextStyle,small: null == small ? _self.small : small // ignore: cast_nullable_to_non_nullable
+as TextStyle,button: null == button ? _self.button : button // ignore: cast_nullable_to_non_nullable
 as TextStyle,
   ));
 }
@@ -73,10 +80,17 @@ as TextStyle,
 
 
 class _WaveTextTheme extends WaveTextTheme {
-  const _WaveTextTheme({required this.body}): super._();
+   _WaveTextTheme({required this.h1, required this.h2, required this.h3, required this.h4, required this.large, required this.body, required this.small, required this.button}): super._();
   
 
+@override final  TextStyle h1;
+@override final  TextStyle h2;
+@override final  TextStyle h3;
+@override final  TextStyle h4;
+@override final  TextStyle large;
 @override final  TextStyle body;
+@override final  TextStyle small;
+@override final  TextStyle button;
 
 /// Create a copy of WaveTextTheme
 /// with the given fields replaced by the non-null parameter values.
@@ -88,12 +102,12 @@ _$WaveTextThemeCopyWith<_WaveTextTheme> get copyWith => __$WaveTextThemeCopyWith
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _WaveTextTheme&&(identical(other.body, body) || other.body == body));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _WaveTextTheme&&(identical(other.h1, h1) || other.h1 == h1)&&(identical(other.h2, h2) || other.h2 == h2)&&(identical(other.h3, h3) || other.h3 == h3)&&(identical(other.h4, h4) || other.h4 == h4)&&(identical(other.large, large) || other.large == large)&&(identical(other.body, body) || other.body == body)&&(identical(other.small, small) || other.small == small)&&(identical(other.button, button) || other.button == button));
 }
 
 
 @override
-int get hashCode => Object.hash(runtimeType,body);
+int get hashCode => Object.hash(runtimeType,h1,h2,h3,h4,large,body,small,button);
 
 
 
@@ -104,7 +118,7 @@ abstract mixin class _$WaveTextThemeCopyWith<$Res> implements $WaveTextThemeCopy
   factory _$WaveTextThemeCopyWith(_WaveTextTheme value, $Res Function(_WaveTextTheme) _then) = __$WaveTextThemeCopyWithImpl;
 @override @useResult
 $Res call({
- TextStyle body
+ TextStyle h1, TextStyle h2, TextStyle h3, TextStyle h4, TextStyle large, TextStyle body, TextStyle small, TextStyle button
 });
 
 
@@ -121,9 +135,16 @@ class __$WaveTextThemeCopyWithImpl<$Res>
 
 /// Create a copy of WaveTextTheme
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? body = null,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? h1 = null,Object? h2 = null,Object? h3 = null,Object? h4 = null,Object? large = null,Object? body = null,Object? small = null,Object? button = null,}) {
   return _then(_WaveTextTheme(
-body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+h1: null == h1 ? _self.h1 : h1 // ignore: cast_nullable_to_non_nullable
+as TextStyle,h2: null == h2 ? _self.h2 : h2 // ignore: cast_nullable_to_non_nullable
+as TextStyle,h3: null == h3 ? _self.h3 : h3 // ignore: cast_nullable_to_non_nullable
+as TextStyle,h4: null == h4 ? _self.h4 : h4 // ignore: cast_nullable_to_non_nullable
+as TextStyle,large: null == large ? _self.large : large // ignore: cast_nullable_to_non_nullable
+as TextStyle,body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+as TextStyle,small: null == small ? _self.small : small // ignore: cast_nullable_to_non_nullable
+as TextStyle,button: null == button ? _self.button : button // ignore: cast_nullable_to_non_nullable
 as TextStyle,
   ));
 }